# ent/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Ent Schema API",
    "description": "This is an auto generated API description made out of an Ent schema definition",
    "version": "0.1.0"
  },
  "paths": {
    "/cards": {
      "get": {
        "tags": [
          "Card"
        ],
        "summary": "List Cards",
        "description": "List Cards.",
        "operationId": "listCard",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "what page to render",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          },
          {
            "name": "itemsPerPage",
            "in": "query",
            "description": "item count to render per page",
            "schema": {
              "type": "integer",
              "maximum": 5000,
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "result Card list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CardList"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "409": {
            "$ref": "#/components/responses/409"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      },
      "post": {
        "tags": [
          "Card"
        ],
        "summary": "Create a new Card",
        "description": "Creates a new Card and persists it to storage.",
        "operationId": "createCard",
        "requestBody": {
          "description": "Card to create",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "password": {
                    "type": "string"
                  },
                  "card": {
                    "type": "integer"
                  },
                  "skill": {
                    "type": "string"
                  },
                  "status": {
                    "type": "string"
                  },
                  "token": {
                    "type": "string"
                  },
                  "cp": {
                    "type": "integer"
                  },
                  "url": {
                    "type": "string"
                  },
                  "count": {
                    "type": "integer"
                  },
                  "author": {
                    "type": "string"
                  },
                  "created_at": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "owner": {
                    "type": "integer"
                  }
                },
                "required": [
                  "password",
                  "owner"
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Card created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CardCreate"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "409": {
            "$ref": "#/components/responses/409"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
    "/cards/{id}": {
      "get": {
        "tags": [
          "Card"
        ],
        "summary": "Find a Card by ID",
        "description": "Finds the Card with the requested ID and returns it.",
        "operationId": "readCard",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the Card",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Card with requested ID was found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CardRead"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "409": {
            "$ref": "#/components/responses/409"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      },
      "delete": {
        "tags": [
          "Card"
        ],
        "summary": "Deletes a Card by ID",
        "description": "Deletes the Card with the requested ID.",
        "operationId": "deleteCard",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the Card",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Card with requested ID was deleted"
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "409": {
            "$ref": "#/components/responses/409"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      },
      "patch": {
        "tags": [
          "Card"
        ],
        "summary": "Updates a Card",
        "description": "Updates a Card and persists changes to storage.",
        "operationId": "updateCard",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the Card",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "description": "Card properties to update",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "card": {
                    "type": "integer"
                  },
                  "skill": {
                    "type": "string"
                  },
                  "status": {
                    "type": "string"
                  },
                  "token": {
                    "type": "string"
                  },
                  "cp": {
                    "type": "integer"
                  },
                  "url": {
                    "type": "string"
                  },
                  "count": {
                    "type": "integer"
                  },
                  "author": {
                    "type": "string"
                  },
                  "owner": {
                    "type": "integer"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Card updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CardUpdate"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "409": {
            "$ref": "#/components/responses/409"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
    "/cards/{id}/d": {
      "description": "Start an draw as done",
      "put": {
        "tags": [
          "Card"
        ],
        "summary": "Draws a card item as done.",
        "operationId": "drawDone",
        "responses": {
          "204": {
            "description": "Item marked as done"
          }
        }
      },
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "integer"
          }
        }
      ]
    },
    "/cards/{id}/owner": {
      "get": {
        "tags": [
          "Card"
        ],
        "summary": "Find the attached User",
        "description": "Find the attached User of the Card with the given ID",
        "operationId": "readCardOwner",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the Card",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User attached to Card with requested ID was found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Card_OwnerRead"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "409": {
            "$ref": "#/components/responses/409"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
...
    "/groups/{id}/users": {
      "get": {
        "tags": [
          "Group"
        ],
        "summary": "List attached Users",
        "description": "List attached Users.",
        "operationId": "listGroupUsers",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the Group",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "what page to render",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "itemsPerPage",
            "in": "query",
            "description": "item count to render per page",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "result Groups list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Group_UsersList"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "409": {
            "$ref": "#/components/responses/409"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
...
    "/users/{id}/card/start": {
      "description": "Start an draw as done",
      "patch": {
        "tags": [
          "Card"
        ],
        "summary": "Draws a card item as done.",
        "operationId": "drawStart",
        "responses": {
          "204": {
            "description": "Item marked as done"
          }
        }
      },
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "integer"
          }
        }
      ]
    },
...
  },
  "components": {
    "schemas": {
...
      "CardList": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "card": {
            "type": "integer"
          },
          "skill": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "cp": {
            "type": "integer"
          },
          "url": {
            "type": "string"
          },
          "count": {
            "type": "integer"
          },
          "author": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id"
        ]
      },
...
    },
    "responses": {
      "400": {
        "description": "invalid input, data invalid",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "integer"
                },
                "status": {
                  "type": "string"
                },
                "errors": {}
              },
              "required": [
                "code",
                "status"
              ]
            }
          }
        }
      },
...
    }
  }
}